refactor: postList 리팩토링
- postCard Widgets 으로 분리
- postList 필터 로직 Model 로 분리
