+++ apps/web/src/widgets/footerNav/ui/FooterNav.tsx
